stix_shifter_modules/azure_sentinel/requirements.txt
azure-identity==1.10.0
azure-monitor-query==1.0.2
pandas==1.4.2
regex==2022.7.9